requirements.txt
async-timeout
channels
channels-redis
Django==2.0.7
djangorestframework==3.8.2
django-webpack-loader==0.6.0

pip install django==2.0.7
pip install djangorestframework==3.8.2
pip install channels==2.1.2